# demo_package/docs/api_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AITM - AI-Powered Threat Modeler",
    "description": "Automated threat modeling using AI agents and MITRE ATT&CK framework",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/predictions/predict-current-risk": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Predict Current Risk",
        "description": "Predict current risk score based on threat landscape data.",
        "operationId": "predict_current_risk_api_v1_predictions_predict_current_risk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreatDataInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__endpoints__predictions__RiskPredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/predict-future-risk": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Predict Future Risk",
        "description": "Predict future risk progression with scenario analysis.",
        "operationId": "predict_future_risk_api_v1_predictions_predict_future_risk_post",
        "parameters": [
          {
            "name": "days_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days Ahead"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreatDataInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FutureRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/train-models": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Train Prediction Models",
        "description": "Train prediction models with historical data.",
        "operationId": "train_prediction_models_api_v1_predictions_train_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/HistoricalDataPoint"
                },
                "type": "array",
                "title": "Historical Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/analyze-trends": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Analyze Risk Trends",
        "description": "Analyze trends in historical risk data.",
        "operationId": "analyze_risk_trends_api_v1_predictions_analyze_trends_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/HistoricalDataPoint"
                },
                "type": "array",
                "title": "Historical Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__endpoints__predictions__TrendAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/predictions/predict-risk": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Predict Risk Legacy",
        "description": "Legacy endpoint: Predict future risk based on historical data.",
        "operationId": "predict_risk_legacy_api_v1_predictions_predict_risk_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_predict_risk_legacy_api_v1_predictions_predict_risk_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_predict_risk_legacy_api_v1_predictions_predict_risk_post": {
        "properties": {
          "historical_data": {
            "items": {},
            "type": "array",
            "title": "Historical Data"
          }
        },
        "type": "object",
        "required": [
          "historical_data"
        ],
        "title": "Body_predict_risk_legacy_api_v1_predictions_predict_risk_post"
      },
...
      "FutureRiskResponse": {
        "properties": {
          "current_risk": {
            "type": "number",
            "title": "Current Risk"
          },
          "days_ahead": {
            "type": "integer",
            "title": "Days Ahead"
          },
          "scenarios": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scenarios"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "type": "object",
        "required": [
          "current_risk",
          "days_ahead",
          "scenarios",
          "recommendations"
        ],
        "title": "FutureRiskResponse",
        "description": "Response model for future risk predictions"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoricalDataPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO timestamp of the data point"
          },
          "risk_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "description": "Risk score (0-1)"
          },
          "threat_data": {
            "$ref": "#/components/schemas/ThreatDataInput"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "risk_score",
          "threat_data"
        ],
        "title": "HistoricalDataPoint",
        "description": "Historical data point for model training"
      },
...
      "ThreatDataInput": {
        "properties": {
          "days_since_last_analysis": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Days Since Last Analysis",
            "description": "Days since last analysis",
            "default": 0
          },
          "analysis_frequency": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Analysis Frequency",
            "description": "How often analyses are performed",
            "default": 1
          },
          "unique_attack_paths": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unique Attack Paths",
            "description": "Number of unique attack paths identified",
            "default": 0
          },
          "avg_attack_complexity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Avg Attack Complexity",
            "description": "Average complexity of attacks (0-1)",
            "default": 0.5
          },
          "critical_vulnerabilities": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Critical Vulnerabilities",
            "description": "Number of critical vulnerabilities",
            "default": 0
          },
          "public_endpoints": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Public Endpoints",
            "description": "Number of public-facing endpoints",
            "default": 0
          },
          "trust_boundaries": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Trust Boundaries",
            "description": "Number of trust boundaries",
            "default": 1
          },
          "data_sensitivity_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Data Sensitivity Score",
            "description": "Data sensitivity score (0-1)",
            "default": 0.5
          },
          "mitre_technique_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Mitre Technique Count",
            "description": "Number of MITRE ATT&CK techniques identified",
            "default": 0
          },
          "mitre_tactic_coverage": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Mitre Tactic Coverage",
            "description": "Number of MITRE tactics covered",
            "default": 0
          },
          "high_impact_techniques": {
            "type": "integer",
            "minimum": 0.0,
            "title": "High Impact Techniques",
            "description": "Number of high-impact techniques",
            "default": 0
          }
        },
        "type": "object",
        "title": "ThreatDataInput",
        "description": "Input data for threat analysis predictions"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__api__endpoints__predictions__RiskPredictionResponse": {
        "properties": {
          "risk_score": {
            "type": "number",
            "title": "Risk Score",
            "description": "Predicted risk score (0-1)"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "description": "Prediction confidence (0-1)"
          },
          "model_predictions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Model Predictions",
            "description": "Individual model predictions"
          },
          "feature_contributions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Feature Contributions",
            "description": "Feature contribution analysis"
          }
        },
        "type": "object",
        "required": [
          "risk_score",
          "confidence",
          "model_predictions",
          "feature_contributions"
        ],
        "title": "RiskPredictionResponse",
        "description": "Response model for risk predictions"
      },
      "app__api__endpoints__predictions__TrendAnalysisResponse": {
        "properties": {
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "slope": {
            "type": "number",
            "title": "Slope"
          },
          "correlation": {
            "type": "number",
            "title": "Correlation"
          },
          "recent_avg": {
            "type": "number",
            "title": "Recent Avg"
          },
          "historical_avg": {
            "type": "number",
            "title": "Historical Avg"
          },
          "volatility": {
            "type": "number",
            "title": "Volatility"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "data_points": {
            "type": "integer",
            "title": "Data Points"
          }
        },
        "type": "object",
        "required": [
          "trend",
          "slope",
          "correlation",
          "recent_avg",
          "historical_avg",
          "volatility",
          "confidence",
          "data_points"
        ],
        "title": "TrendAnalysisResponse",
        "description": "Response model for trend analysis"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}